requirements.txt
black==22.12.0
click==8.1.3
mypy-extensions==0.4.3
pathspec==0.11.0
platformdirs==2.6.2
pygame==2.1.2
PyYAML==6.0
tomli==2.0.1
